#2 Добавил модель, сервис, контроллер, репозитроий для Product. Изменил тип цены в миграций на float

+++ Avtolavka/core/manager/src/main/resources/db/migration/V1000__2022.01.11_14.00_create_table_products.sql
@@ -2,7 +2,7 @@ create table "products"(
                            "id" SERIAL NOT NULL,
                            "name" varchar NOT NULL ,
                            "quantity" int NOT NULL ,
-                           "purchase_price" numeric CHECK (purchase_price > 0) NOT NULL ,
-                           "selling_price" numeric CHECK (selling_price > 0) NOT NULL ,
+                           "purchase_price" float CHECK (purchase_price > 0) NOT NULL ,
+                           "selling_price" float CHECK (selling_price > 0) NOT NULL ,
                            "in_stock" boolean NOT NULL
 )
